# specs/001-shared-expense-ledger/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FairShareApp Shared Expense Ledger API
  version: 0.1.0
  description: Contract for groups, invites, expenses, settlements, balances, and ledger history.
servers:
  - url: https://api.fairshareapp.local
security:
  - bearerAuth: []
paths:
...
  /groups/{groupId}/invites:
    post:
      summary: Create invite
      operationId: createInvite
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "201":
          description: Invite created
    get:
      summary: List invites
      operationId: listInvites
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Invite list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InviteResponse"
  /groups/{groupId}/invites/{inviteId}/revoke:
    post:
      summary: Revoke invite
      operationId: revokeInvite
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/InviteId"
      responses:
        "200":
          description: Invite revoked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InviteId:
      name: inviteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateInviteRequest:
      type: object
      required: [channel, recipient, expiresAt]
      properties:
        channel:
          type: string
          enum: [Link, Email]
        recipient:
          type: string
        expiresAt:
          type: string
          format: date-time
    InviteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [Pending, Accepted, Revoked, Expired]
        expiresAt:
          type: string
          format: date-time
...